refactor: move abstract classes to parent modules
